requirements.txt
matplotlib==3.8.0
numpy==1.26.2
optuna==3.3.0
pandas==1.5.3
rdkit_pypi==2022.9.4
scikit_learn==1.3.1
setuptools==68.2.2
torch==2.0.1
torch_geometric==2.0.3
